added capaability to give nonlinear heat solve.  tested by comparison to Reddy problem in the nonlinear book.  about to add 2 degrees of freedom to match partial currents at the boundary

+++ HeatSolver/density_corr.f90
@@ -12,11 +12,11 @@ subroutine density_corr(T, density)
 implicit none
 
 ! Dummy
-    real(kind=8), intent(in)  :: T
-    real(kind=8), intent(out) :: density 
+    real, intent(in)  :: T
+    real, intent(out) :: density 
 ! Local
-    real(kind=8) :: a
-    real(kind=8) :: b  
+    real :: a
+    real :: b  
     
     a = 2579.3 
     b = -0.624
